requirements.txt
transformers==4.46.3
datasets==2.19.0
mlflow==2.13.0
seqeval>=1.2.2
boto3>=1.34.0
psycopg2-binary>=2.9.9
scikit-learn>=1.4.0
python-dateutil>=2.9.0
huggingface-hub>=0.22.0
accelerate>=0.29.0
torch>=2.1.0
numpy>=1.26.0
nltk>=3.8.0
pyyaml>=6.0
